+++ src/interpret.rs
@@ -60,9 +60,9 @@ fn configure_dind(
 }
 
 fn configure_floki_user_env(cmd: DockerCommandBuilder, env: &Environment) -> DockerCommandBuilder {
-    let (ref user, ref group) = env.user_details;
-    let new_cmd = cmd.add_environment("FLOKI_HOST_UID", &user);
-    new_cmd.add_environment("FLOKI_HOST_GID", &group)
+    let user = env.user_details;
+    let new_cmd = cmd.add_environment("FLOKI_HOST_UID", user.uid.to_string());
+    new_cmd.add_environment("FLOKI_HOST_GID", user.gid.to_string())
 }
 
 fn configure_floki_host_mountdir_env(
@@ -78,8 +78,9 @@ fn configure_forward_user(
     env: &Environment,
 ) -> DockerCommandBuilder {
     if config.forward_user {
-        let (ref user, ref group) = env.user_details;
-        cmd.add_docker_switch(&format!("--user {}:{}", user, group))
+        let user = env.user_details;
+        cmd.add_docker_switch("--user")
+            .add_docker_switch(&format!("{}:{}", user.uid, user.gid))
     } else {
         cmd
     }
@@ -107,7 +108,9 @@ fn configure_docker_switches(
 ) -> DockerCommandBuilder {
     let mut cmd = cmd;
     for switch in &config.docker_switches {
-        cmd = cmd.add_docker_switch(&switch);
+        for s in switch.split_whitespace() {
+            cmd = cmd.add_docker_switch(s);
+        }
     }
 
     cmd
@@ -118,15 +121,11 @@ fn configure_working_directory(
     env: &Environment,
     config: &FlokiConfig,
 ) -> DockerCommandBuilder {
-    cmd.set_working_directory(
-        get_working_directory(
-            &env.current_directory,
-            &env.floki_root,
-            &path::PathBuf::from(&config.mount),
-        )
-        .to_str()
-        .unwrap(),
-    )
+    cmd.set_working_directory(get_working_directory(
+        &env.current_directory,
+        &env.floki_root,
+        &path::PathBuf::from(&config.mount),
+    ))
 }
 
 /// Add mounts for each of the passed in volumes